cpc -> cpc, curvecontainer, cpcinverter
