--- fastlane_bot/tools/cpcbase.py
@@ -1,2 +1,0 @@
-"""temporary module to maintain the legacy name for the curvebase module"""
-from curvebase import *
